Move Gradle plugin logic to DetektKotlinCompilerPlugin
