+++ src/main/kotlin/io/github/detekt/gradle/DetektKotlinCompilerPlugin.kt
@@ -4,7 +4,9 @@ import io.github.detekt.compiler.plugin.Options
 import io.github.detekt.gradle.extensions.DetektExtension
 import org.gradle.api.Project
 import org.gradle.api.file.ConfigurableFileCollection
+import org.gradle.api.plugins.ReportingBasePlugin
 import org.gradle.api.provider.Provider
+import org.gradle.api.reporting.ReportingExtension
 import org.jetbrains.kotlin.gradle.plugin.KotlinCompilation
 import org.jetbrains.kotlin.gradle.plugin.KotlinCompilerPluginSupportPlugin
 import org.jetbrains.kotlin.gradle.plugin.KotlinPlatformType
@@ -20,6 +22,16 @@ class DetektKotlinCompilerPlugin : KotlinCompilerPluginSupportPlugin {
 
     override fun apply(target: Project) {
         project = target
+
+        target.pluginManager.apply(ReportingBasePlugin::class.java)
+        val extension = target.extensions.create(DETEKT_NAME, DetektExtension::class.java, target)
+        extension.reportsDir = target.extensions.getByType(ReportingExtension::class.java).file(DETEKT_NAME)
+
+        val defaultConfigFile = getDefaultConfigFile(target)
+
+        if (defaultConfigFile.exists()) {
+            extension.config = target.files(defaultConfigFile)
+        }
     }
 
     override fun applyToCompilation(kotlinCompilation: KotlinCompilation<*>): Provider<List<SubpluginOption>> {
@@ -49,6 +61,9 @@ class DetektKotlinCompilerPlugin : KotlinCompilerPluginSupportPlugin {
 
     override fun isApplicable(kotlinCompilation: KotlinCompilation<*>): Boolean =
         kotlinCompilation.platformType in setOf(KotlinPlatformType.jvm, KotlinPlatformType.androidJvm)
+
+    private fun getDefaultConfigFile(target: Project) =
+        target.file("${target.rootProject.layout.projectDirectory.dir(CONFIG_DIR_NAME)}/$CONFIG_FILE")
 }
 
 internal fun ConfigurableFileCollection.toDigest(): String {
